+++ python/phenomena/particles/particle_boosted.py
@@ -7,7 +7,7 @@ from particle import Particle, ParticleDT, toDictionary
 
 from phenomena.particles.kinematics.decay.calculations import DecayCalc
 from phenomena.particles.kinematics.parameters import boostParams
-from phenomena.particles.virtual_particle import VirtualChannel
+from phenomena.particles.particle_virtual import VirtualChannel
 
 #Santi was here
